requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
httpx>=0.23.0
pyyaml>=6.0.1
python-json-logger>=2.0.7
structlog>=23.1.0  # For structured logging
python-multipart
sentence-transformers
faiss-cpu
slowapi
jinja2>=3.1.2  # For template rendering